Lucene.Net.TestFramework: Added overloads for int to Assert.AreEqual (addresses #295, #261)

+++ src/Lucene.Net.Tests/Util/Fst/TestFSTs.cs
@@ -1882,7 +1882,7 @@ namespace Lucene.Net.Util.Fst
                 Assert.AreEqual(result.Bytes[result.Offset], arc);
                 for (int byteIDX = 1; byteIDX < result.Length; byteIDX++)
                 {
-                    Assert.AreEqual(0, result.Bytes[result.Offset + byteIDX]);
+                    Assert.AreEqual((byte)0, result.Bytes[result.Offset + byteIDX]);
                 }
             }
         }
